fix: Resolve MSVC type mismatch errors in Windows build
- TuningPlayer.cpp: Cast 3 to size_t in std::min() call to match midiNotes.size() type
- AYPlugin.cpp: Cast sampleRate (double) to int in AyumiEmulator constructor call

These fixes address MSVC's strict type checking for std::min and constructor arguments.

+++ src/plugins/uZX/aychip/AYPlugin.cpp
@@ -121,7 +121,7 @@ void AYChipPlugin::reset() {
     const int ayumiMode = 3;
 
     if (chip == nullptr) {
-        chip = std::make_unique<AyumiEmulator>(sampleRate, staticParams.chipClock.getStoredValue() * MHz,
+        chip = std::make_unique<AyumiEmulator>(static_cast<int>(sampleRate), staticParams.chipClock.getStoredValue() * MHz,
                                                staticParams.chipType.getStoredValue(),
                                                ayumiMode);
     } else {
